requirements.txt
pyspark
boto3
requests
python-dotenv
delta-spark